Validator: undefined, charset, explode
The rule *undefined* states, that item at given position must not
exists. It useful to state, that some key of the object is not included.
It works same way has "optional", however, optional should be used along
with other rule expressing that value is optional

The charset [...] states, that item is string having characters from
given set of chars, similar to regexp. Ranges are also supported [a-z].
You can also ask for [ and ]. If you need to test for -, specify it as
the first or as the last or between ranges [a-z-0-9]. Begin with ^ to
negate set. All characters are evaluated as unicode.

explode splits string into an array, where each item can be evaluated
separatedly. ["explode","separator",rule, limit] - limit is optional